AgTextLabel-class {Rgraphviz}	R Documentation

Class "AgTextLabel": Represents a graphviz text label

Description

This class is used to represent the textlabel object in C from the Graphviz library

Objects from the Class

Objects can be created by calls of the form new("AgTextLabel", ...).

Slots

labelText:

Object of class "character" The actual label text

labelLoc:

Object of class "xyPoint" The location of the label

labelJust:

Object of class "character" The justification of the label

labelWidth:

Object of class "integer" The width of the label

labelColor:

Object of class "character" The color to print the label

labelFontsize:

Object of class "numeric" The font size for the label

Methods

labelText

Retrieves the labelText slot

labelLoc

Retrieves the labelLoc slot

labelJust

Retrieves the labelJust slot

labelWidth

Retrieves the labelWidth slot

labelColor

Retrieves the labelColor slot

labelFontsize

Retrieves the labelFontsize slot

Author(s)

Jeff Gentry

See Also

AgEdge-class, AgNode-class

Examples

V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)

## Make the labels be the edge weights.  This code is from the vignette
eAttrs <- list()
ew <- edgeWeights(g1)
lw <- unlist(unlist(ew))
toRemove <- removedEdges(g1)
lw <- lw[-toRemove]
names(lw) <- edgeNames(g1)
eAttrs$label <- lw

z <- agopen(g1,"foo", edgeAttrs=eAttrs)
x <- AgEdge(z)  ## list of AgEdge objects
x[[1]]  ## AgEdge
a <- txtLabel(x[[1]])
a ## txtLabel object

labelText(a)
labelLoc(a)
labelJust(a)
labelWidth(a)
labelColor(a)
labelFontsize(a)

[Package Rgraphviz version 2.26.0 Index]