BezierCurve-class {Rgraphviz}	R Documentation

Class "BezierCurve": A class to describe a Bezier curve

Description

This class is used to represent a Bezier curve in R, which can then be used for other applications, plotted, etc

Objects from the Class

Objects can be created by calls of the form new("BezierCurve", ...).

Slots

cPoints:

Object of class "list": A list of xyPoint objects, representing control points for the curve

Methods

cPoints:

Returns the cPoints slot

pointList:

Returns a list of all points having been processed with teh getPoints method of xyPoint

bezierPoints:

Returns a matrix giving x & y points (by column) for the complete Bezier curve

lines:

Draws the Bezier curve

show:

Provides a concise display of information

Author(s)

Jeff Gentry

See Also

xyPoint

Examples

V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)
z <- agopen(g1,"foo")
x <- AgEdge(z)  ## list of AgEdge objects
x[[1]]  ## AgEdge
a <- splines(x[[1]])[[1]]
a  ## BezierCurve

cPoints(a)
pointList(a)
bezierPoints(a)

[Package Rgraphviz version 2.26.0 Index]