agwrite {Rgraphviz}	R Documentation

Read / write Ragraph objects

Description

These functions will write an Ragraph object to or from a file.

Usage

agwrite(graph, filename)
agread(filename, layoutType="dot", layout=TRUE)

Arguments

graph
An object of class Ragraph

filename
The input or output filename

layoutType
character(1) specifying the format of the input file. must be one of graphvizCapabilities()$layoutTypes.

layout
logical(1) indicating whether graphLayout is to be called on the result of file input.

Details

These function are wrappers to agwrite() and agread() calls in Graphviz.

Author(s)

Jeff Gentry

See Also

agopen, agread

Examples

V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)
z <- agopen(g1, "foo", layout=FALSE) ## default layoutType 'dot'
file <- tempfile()
agwrite(z, file)

## supported input types
graphvizCapabilities()$layoutTypes
g2 <- agread(file, layout=TRUE)
if (interactive())
   plot(g2)

[Package Rgraphviz version 2.26.0 Index]