boundingBox-class {Rgraphviz}	R Documentation

Class "boundingBox": A class to describe the bounding box of a Ragraph

Description

The boundingBox class is used to describe the dimensons of the bounding box for a laid out Ragraph

Objects from the Class

Objects can be created by calls of the form new("boundingBox", ...).

Slots

botLeft:

Object of class "xyPoint" Defines the bottom left point of the bounding box

upRight:

Object of class "xyPoint" Defines the upper right point of the bounding box

Methods

botLeft

Retrieve the botLeft slot

upRight

Retrieve the upRight slot

Author(s)

Jeff Gentry

See Also

Ragraph, graph-class

Examples

V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)
z <- agopen(g1,"foo")
x <- boundBox(z)
x

botLeft(x)
upRight(x)

[Package Rgraphviz version 2.26.0 Index]