getDefaultAttrs {Rgraphviz}	R Documentation

Functions to generate and check global attribute lists

Description

The getDefaultAttrs function can be used to generate a default global attribute list for Graphviz. The checkAttrs function can be used to verify that such a list is valid for use.

Usage

getDefaultAttrs(curAttrs = list(), layoutType = "dot")
checkAttrs(attrs)

Arguments

curAttrs
Any attributes currently defined

layoutType
The layout method being used. Defaults to dot, and see graphvizCapabilities()$layoutTypes for possible values.

attrs
An attribute list of graphviz attributes

Details

The getDefaultAttrs function generates a four element list (elements being “graph”, “cluster”, “node” and “edge”). Contained in each is another list where the element names correspond to attributes and the value is the value for that attribute. This list can be used to set global attributes in Graphviz, and the exact list returned by getDefaultAttrs represents the values that are used as basic defaults.

The checkAttrs function can be used to verify that a global attribute list is properly formed.

Author(s)

Jeff Gentry

See Also

agopen, plot.graph

Examples

  z <- getDefaultAttrs()
  checkAttrs(z)

[Package Rgraphviz version 2.26.0 Index]