graphData-methods {Rgraphviz}	R Documentation

Get and set attributes of an Ragraph object

Description

Attributes of a graph can be accessed using graphData. The attributes could be defined using graphDataDefaults.

Usage

graphData(self, attr)
graphData(self, attr) <- value

Arguments

self
A Ragraph-class instance

attr
A character vector of length one specifying the name of a graph attribute

value
A character vector to store as the attribute values

Author(s)

Li Long <li.long@isb-sib.ch>

[Package Rgraphviz version 2.26.0 Index]