graphDataDefaults-methods {Rgraphviz}	R Documentation

Get and set default attributes for an Ragraph

Description

Get/Set default values for attributes associated with a graph.

Usage

graphDataDefaults(self)
graphDataDefaults(self, attr) <- value

Arguments

self
A Ragraph-class instance

attr
A character value giving the name of the attribute

value
A character value as the default value for the specified attribute

Author(s)

Li Long <li.long@isb-sib.ch>

[Package Rgraphviz version 2.26.0 Index]