graphLayout {Rgraphviz}	R Documentation

A function to layout graph locations

Description

This function will take an object of class Ragraph and will perform a libgraph layout on the graph locations.

Usage

graphLayout(graph, layoutType=graph@layoutType)

Arguments

graph
An object of type Ragraph

layoutType
layout algorithm to use

Details

If the graph has already been laid out, this function merely returns its parameter. Otherwise, it will perform a libgraph layout and retrieve the appropriate location information.

Value

A laid out object of type Ragraph.

Author(s)

Jeff Gentry

See Also

agopen

Examples

V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)
z <- agopen(g1,"foo",layout=FALSE)
x <- z
a <- graphLayout(z)

[Package Rgraphviz version 2.26.0 Index]