graphvizVersion {Rgraphviz}	R Documentation

A function to determine graphviz library version

Description

This function will query the graphviz libraries that the package was built against and report what version of Graphviz is being used.

Usage

graphvizVersion()

Value

A list with three elements, two of class numeric_version. The first element named installed_version represents the version of Graphviz that is being used by the package. The second element named build_version represents the version of Graphviz that was used to build the package. A mismatch between these two versions may indicate problems. The third element is a logical named bundled_graphviz and indicates if Rgraphviz is using the bundled Graphviz (default) or an external Graphviz.

Author(s)

Jeff Gentry, modified by Kasper Daniel Hansen

Examples

graphvizVersion()

[Package Rgraphviz version 2.26.0 Index]