makeNodeAttrs {Rgraphviz}	R Documentation

make a list of character vectors that can be used as a value for the nodeAttrs argument in agopen

Description

make a list of character vectors that can be used as a value for the nodeAttrs argument in agopen

Usage

makeNodeAttrs(g, label = nodes(g), shape = "ellipse",
              fillcolor = "#e0e0e0", ...)

Arguments

g
graph

label
character of length either 1 or numnodes(g). If the length is 1, the value is recycled.

shape
character of length either 1 or numnodes(g)

fillcolor
character of length either 1 or numnodes(g)

...
further named arguments that are character vectors of length either 1 or numNodes(g)

Details

This function is trivial but convenient.

Value

A list of named character vectors, each of which with length numNodes(g).

Author(s)

Wolfgang Huber <huber@ebi.ac.uk>

Examples

g <- randomEGraph(letters[1:10], p=0.2)
makeNodeAttrs(g)

[Package Rgraphviz version 2.26.0 Index]