pieGlyph {Rgraphviz}	R Documentation

A function to plot pie graphs as a glyph

Description

This function allows the user to plot a pie graph at a specified x/y location in a plotting region.

Usage

pieGlyph(x, xpos, ypos, labels = names(x), edges = 200, radius = 0.8, density = NULL, angle = 45, col = NULL, border = NULL, lty = NULL, main = NULL, ...)

Arguments

xpos
The x location of the glyph

ypos
The Y location of the glyph

x
a vector of positive quantities. The values in x are displayed as the areas of pie slices.

labels
a vector of character strings giving names for the slices. For empty or NA labels, no pointing line is drawn either.

edges
the circular outline of the pie is approximated by a polygon with this many edges.

radius
the pie is drawn centered in a square box whose sides range from -1 to 1. If the character strings labeling the slices are long it may be necessary to use a smaller radius.

density
the density of shading lines, in lines per inch. The default value of NULL means that no shading lines are drawn. Non-positive values of density also inhibit the drawing of shading lines.

angle
the slope of shading lines, given as an angle in degrees (counter-clockwise).

col
a vector of colors to be used in filling or shading the slices. If missing a set of 6 pastel colours is used, unless density is specified when par("fg") is used.

border, lty
(possibly vectors) arguments passed to polygon which draws each slice.

main
an overall title for the plot.

...
graphical parameters can be given as arguments to pie. They will affect the main title and labels only.

Author(s)

R. Gentleman, F. Sim

See Also

pie

Examples

   plot(1:10, col="white")
   pieGlyph(1:20, 5, 5)

[Package Rgraphviz version 2.26.0 Index]