removedEdges {Rgraphviz}	R Documentation

A Function To List Removed Edges

Description

This function can be used to retrieve a numerical vector which will describe which edges in a graph would be removed if recipEdges is set to combined during plotting.

Usage

removedEdges(graph)

Arguments

graph
An object of class graph or Ragraph, the graph to perform this operation on

Details

This function will simply detect which (if any) edges in a graph would be removed during combination of reciprocated edges.

Value

A numerical vector, where the values correspond to removed edges.

Author(s)

Jeff Gentry

See Also

edgeNames, agopen, buildEdgeList

Examples

set.seed(123)
V <- letters[1:10]
M <- 1:4
g1 <- randomGraph(V, M, .2)
removedEdges(g1)

[Package Rgraphviz version 2.26.0 Index]