xyPoint-class {Rgraphviz}	R Documentation

Class "xyPoint": A class to represent a X/Y coordinate.

Description

This class is used to describe a coordinate in 2-dimensional (X/Y) space

Objects from the Class

Objects can be created by calls of the form new("xyPoint", ...).

Slots

x:

Object of class "numeric" The x coordinate

y:

Object of class "numeric" The y coordinate

Methods

getX

Returns the value stored in the x slot

getY

Returns the value stored in the y slot

getPoints

Returns a vector of two numerical values representing the x and y positions

show

Display information about the object in a concise fashion

Author(s)

Jeff Gentry

Examples

   z <- new("xyPoint", x=150, y=30)
   z
   getPoints(z)

   getX(z)
   getY(z)

[Package Rgraphviz version 2.26.0 Index]